requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
dj-database-url==0.5.0
Django==3.0.8
django-filter==2.3.0
django-heroku==0.3.1
django-postgres-copy==2.4.2
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.10
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
reportlab==3.5.48
requests==2.24.0
selenium==3.141.0
sqlparse==0.3.1
urllib3==1.25.10
webdriver-manager==3.2.1
whitenoise==5.2.0
